pix pr and refactor
